+++ crates/apollo-compiler/tests/validation/mod.rs
@@ -5,6 +5,7 @@ mod variable;
 
 use apollo_compiler::ast;
 use apollo_compiler::ExecutableDocument;
+use apollo_compiler::GraphQLLocation;
 use apollo_compiler::Schema;
 
 #[test]
@@ -45,18 +46,35 @@ query {
 }
 "#;
 
+    let json = expect_test::expect![[r#"
+{
+  "message": "an executable document must not contain an object type definition",
+  "locations": [
+    {
+      "line": 2,
+      "column": 1
+    }
+  ]
+}"#]];
+
     let schema = Schema::parse(input_type_system, "schema.graphql");
     let executable = ExecutableDocument::parse(&schema, input_executable, "query.graphql");
 
     schema.validate().unwrap();
-    let errors = executable
-        .validate(&schema)
-        .unwrap_err()
-        .to_string_no_color();
+    let diagnostics = executable.validate(&schema).unwrap_err();
+    let errors = diagnostics.to_string_no_color();
     assert!(
         errors.contains("an executable document must not contain an object type definition"),
         "{errors}"
     );
+
+    diagnostics.iter().for_each(|diag| {
+        assert_eq!(
+            diag.get_line_column(),
+            Some(GraphQLLocation { line: 2, column: 1 })
+        );
+        json.assert_eq(&serde_json::to_string_pretty(&diag.to_json()).unwrap());
+    });
 }
 
 #[test]
@@ -113,19 +131,34 @@ fragment q on TestObject {
     ...q
 }
 "#;
+    let json = expect_test::expect![[r#"
+{
+  "message": "compiler error: `q` fragment cannot reference itself",
+  "locations": [
+    {
+      "line": 8,
+      "column": 1
+    }
+  ]
+}"#]];
 
     let schema = Schema::parse(input_type_system, "schema.graphql");
     let executable = ExecutableDocument::parse(&schema, input_executable, "query.graphql");
 
     schema.validate().unwrap();
-    let errors = executable
-        .validate(&schema)
-        .unwrap_err()
-        .to_string_no_color();
+    let diagnostics = executable.validate(&schema).unwrap_err();
+    let errors = diagnostics.to_string_no_color();
     assert!(
         errors.contains("`q` fragment cannot reference itself"),
         "{errors}"
     );
+    diagnostics.iter().for_each(|diag| {
+        assert_eq!(
+            diag.get_line_column(),
+            Some(GraphQLLocation { line: 8, column: 1 })
+        );
+        json.assert_eq(&serde_json::to_string_pretty(&diag.to_json()).unwrap());
+    });
 }
 
 #[test]
@@ -141,14 +174,32 @@ fn validation_without_type_system() {
         "#,
         "dupe_frag.graphql",
     );
-    let errors = doc
-        .validate_standalone_executable()
-        .unwrap_err()
-        .to_string_no_color();
+    let json = expect_test::expect![[r#"
+{
+  "message": "compiler error: fragment `A` must be used in an operation",
+  "locations": [
+    {
+      "line": 2,
+      "column": 13
+    }
+  ]
+}"#]];
+    let diagnostics = doc.validate_standalone_executable().unwrap_err();
+    let errors = diagnostics.to_string_no_color();
     assert!(
         errors.contains("fragment `A` must be used in an operation"),
         "{errors}"
     );
+    diagnostics.iter().for_each(|diag| {
+        assert_eq!(
+            diag.get_line_column(),
+            Some(GraphQLLocation {
+                line: 2,
+                column: 13
+            })
+        );
+        json.assert_eq(&serde_json::to_string_pretty(&diag.to_json()).unwrap());
+    });
 
     let doc = ast::Document::parse(
         r#"
@@ -158,24 +209,57 @@ fn validation_without_type_system() {
         "#,
         "dupe_frag.graphql",
     );
-    let errors = doc
-        .validate_standalone_executable()
-        .unwrap_err()
-        .to_string_no_color();
+    let json = expect_test::expect![[r#"
+{
+  "message": "the fragment `A` is defined multiple times in the document",
+  "locations": [
+    {
+      "line": 3,
+      "column": 22
+    }
+  ]
+}"#]];
+    let diagnostics = doc.validate_standalone_executable().unwrap_err();
+    let errors = diagnostics.to_string_no_color();
     assert!(
         errors.contains("the fragment `A` is defined multiple times in the document"),
         "{errors}"
     );
+    diagnostics.iter().for_each(|diag| {
+        assert_eq!(
+            diag.get_line_column(),
+            Some(GraphQLLocation {
+                line: 3,
+                column: 22
+            })
+        );
+        json.assert_eq(&serde_json::to_string_pretty(&diag.to_json()).unwrap());
+    });
 
     let doc = ast::Document::parse(r#"{ ...A }"#, "unknown_frag.graphql");
-    let errors = doc
-        .validate_standalone_executable()
-        .unwrap_err()
-        .to_string_no_color();
+    let json = expect_test::expect![[r#"
+{
+  "message": "compiler error: cannot find fragment `A` in this document",
+  "locations": [
+    {
+      "line": 1,
+      "column": 3
+    }
+  ]
+}"#]];
+    let diagnostics = doc.validate_standalone_executable().unwrap_err();
+    let errors = diagnostics.to_string_no_color();
     assert!(
         errors.contains("cannot find fragment `A` in this document"),
         "{errors}"
     );
+    diagnostics.iter().for_each(|diag| {
+        assert_eq!(
+            diag.get_line_column(),
+            Some(GraphQLLocation { line: 1, column: 3 })
+        );
+        json.assert_eq(&serde_json::to_string_pretty(&diag.to_json()).unwrap());
+    });
 }
 
 #[test]